Split RoleUpdater into implementation and interface
